requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
Pillow==11.2.1
numpy==2.2.4
gunicorn==23.0.0
